requirements.txt
pandas==2.2.2
langchain==0.2.16
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4